duckdb-sql-editor-refactored/requirements.txt
fasthtml>=0.1.0
monsterui>=0.1.0
duckdb>=0.9.0
python-dotenv>=1.0.0
requests>=2.31.0 